chore(cleanup) move components into the components folder
