+++ src/app/components/clipboard.tsx
@@ -1,8 +1,8 @@
 import React = require("react");
 import Radium = require('radium');
 import csx = require('csx');
-import {BaseComponent} from "./ui";
-import * as ui from "./ui";
+import {BaseComponent} from "../ui";
+import * as ui from "../ui";
 
 export interface Props {
     text: string;
