+++ src/app/components/icon.tsx
@@ -1,9 +1,9 @@
 import React = require("react");
 import Radium = require('radium');
 import csx = require('csx');
-import {BaseComponent} from "./ui";
-import * as ui from "./ui";
-import * as pure from "../common/pure";
+import {BaseComponent} from "../ui";
+import * as ui from "../ui";
+import * as pure from "../../common/pure";
 
 export interface Props extends React.HTMLAttributes {
     name: string;
