+++ src/app/components/robocop.tsx
@@ -3,11 +3,11 @@ require('./robocop.css');
 import React = require("react");
 import Radium = require('radium');
 import csx = require('csx');
-import {BaseComponent} from "./ui";
-import * as ui from "./ui";
-import * as utils from "../common/utils";
-import * as styles from "./styles/styles";
-import * as state from "./state/state";
+import {BaseComponent} from "../ui";
+import * as ui from "../ui";
+import * as utils from "../../common/utils";
+import * as styles from "../styles/styles";
+import * as state from "../state/state";
 
 export interface Props {
 }
