+++ src/app/fileTree.tsx
@@ -11,14 +11,14 @@ import * as styles from "./styles/styles";
 import * as state from "./state/state";
 import {connect} from "react-redux";
 import {StoreState} from "./state/state";
-import {Icon} from "./icon";
+import {Icon} from "./components/icon";
 import * as commands from "./commands/commands";
 let {DraggableCore} = ui;
 import {getDirectory,getFileName} from "../common/utils";
-import {Robocop} from "./robocop";
+import {Robocop} from "./components/robocop";
 import {inputDialog} from "./dialogs/inputDialog";
 import * as Mousetrap from "mousetrap";
-import * as clipboard from "./clipboard";
+import * as clipboard from "./components/clipboard";
 import * as pure from "../common/pure";
 import {tabState} from "./tabs/v2/appTabsContainer";
 type TruthTable = utils.TruthTable;
