+++ src/app/mainPanel.tsx
@@ -10,9 +10,9 @@ import {cast,server} from "../socket/socketClient";
 import * as commands from "./commands/commands";
 import * as types from "../common/types";
 import {AvailableProjectConfig} from "../common/types";
-import {Clipboard} from "./clipboard";
+import {Clipboard} from "./components/clipboard";
 import {PendingRequestsIndicator} from "./pendingRequestsIndicator";
-import {Icon} from "./icon";
+import {Icon} from "./components/icon";
 import {ButtonBlack} from "./components/buttons";
 import {InputBlack} from "./components/inputs";
 let {DraggableCore} = ui;
