+++ src/app/omniSearch/omniSearch.tsx
@@ -10,13 +10,13 @@ import {debounce,createMap,rangeLimited,getFileName} from "../../common/utils";
 import {cast, server, Types} from "../../socket/socketClient";
 import * as commands from "../commands/commands";
 import {match, filter as fuzzyFilter} from "fuzzaldrin";
-import {Icon} from "../icon";
+import {Icon} from "../components/icon";
 import {TypedEvent} from "../../common/events";
 import * as state from "../state/state";
 import * as types from "../../common/types";
 import {AvailableProjectConfig} from "../../common/types";
 import * as CodeMirror from "codemirror";
-import {Robocop} from "../robocop";
+import {Robocop} from "../components/robocop";
 import * as utils from "../../common/utils";
 
 /** Stuff shared by the select list view */
