+++ src/app/tabs/v2/astView.tsx
@@ -10,7 +10,7 @@ import * as $ from "jquery";
 import * as styles from "../../styles/styles";
 import * as onresize from "onresize";
 import {ErrorMessage} from "../errorMessage";
-import {Clipboard} from "../../clipboard";
+import {Clipboard} from "../../components/clipboard";
 
 import {Types} from "../../../socket/socketContract";
